+++ class/actions_advancedproductsearch.class.php
@@ -104,7 +104,7 @@ class ActionsAdvancedProductSearch extends \advancedproductsearch\RetroCompatCom
 			/** @var CommonObject $object */
 
 			// STATUS DRAFT ONLY AND NOT IN EDIT MODE
-		    if(!empty($object->statut) || $action=='editline'){
+		    if(!empty($object->statut) || !empty($object->status)  || $action=='editline'){
 		        return 0;
 		    }
 
@@ -146,7 +146,8 @@ class ActionsAdvancedProductSearch extends \advancedproductsearch\RetroCompatCom
 				'MAIN_MAX_DECIMALS_UNIT' => getDolGlobalString('MAIN_MAX_DECIMALS_UNIT') ,
 				'interface_url' => dol_buildpath('advancedproductsearch/scripts/interface.php',1),
 				'js_url' => dol_buildpath('advancedproductsearch/js/advancedproductsearch.js',1),
-				'supplierElements' => $AdvancedProductSearch->supplierElements
+				'supplierElements' => $AdvancedProductSearch->supplierElements,
+				'displayResultsOnOpen' => getDolGlobalInt('APS_NO_DISPLAY_RESULTS_ON_OPEN') ? 0 : 1
 			);
 
 			?>